+++ src/Core/Movies.Application/DependencyInjection.cs
@@ -1,0 +1,14 @@
+using Microsoft.Extensions.DependencyInjection;
+using Movies.Application.Abstractions;
+
+namespace Movies.Application;
+
+public static class DependencyInjection
+{
+    public static IServiceCollection AddApplication(this IServiceCollection services)
+    {
+        services.AddScoped<IMovieService, MovieService>();
+
+        return services;
+    }
+}